backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.5.2",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-openai>=1.1.12",
    "pydantic-settings>=2.13.1",
    "python-pptx>=1.0.2",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "email-validator>=2.2.0",
    "psycopg2-binary>=2.9.10",
    "httpx>=0.28.1",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.2",
    "locust>=2.43.4",
    "boto3>=1.42.91",
    "langgraph>=1.1.6",
    "cairosvg>=2.9.0",
    "pillow>=12.2.0",
    "openai>=2.31.0",
]
